+++ src/app.scss
@@ -9,7 +9,7 @@
     no-repeat;
 }
 
-.main {
+main {
   padding: var(--header-height);
   min-height: 100vh;
 }
